back-end/auth-api/pyproject.toml
[tool.poetry]
name = "auth-api"
version = "0.1.0"
description = "auth api love you"
authors = ["hyeonproject"]
license = "free"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
hypercorn = {extras = ["uvloop"], version = "^0.13.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy = "^1.4.36"
asyncpg = "^0.25.0"
pytest = "^7.1.2"
httpx = "^0.22.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.111.2"}
google-api-python-client = "^2.47.0"
update = "^0.0.1"
google-auth = "^2.6.6"
google-auth-oauthlib = "^0.5.1"
google-auth-httplib2 = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
